update Python connector license in pyproject toml file (#68153)
## What
In the migration to update connector open source licenses, we missed a
couple references. This change necessary to pass our pre-release checks
since we assert that the license in the connector's pyproject.toml file
matches the license in the connector's metadata file. I don't believe
this warrants bumping connector versions.

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [x] YES 💚
- [ ] NO ❌