chore: Allow for setting control version on github action (#68110)
## What
<!--
* Describe what the change is solving. Link all GitHub issues related to
this change.
-->
In some cases (recent release or progressive rollout), there were no
syncs with the most recent version of the connector. We still want to be
able to run regression tests.

## How
<!--
* Describe how code changes achieve the solution.
-->
We will allow this by adding an optional input in the GitHub action. 

Other solutions that were discarded:
* Find the version released before the one active in prod. This is not
easily possible as we use [the cloud
registry](https://storage.googleapis.com/prod-airbyte-cloud-connector-metadata-service/registries/v0/cloud_registry.json)
in order to be able to know the version and there is only the most
recent version
* Try to get the previous version based on the version there was an
attempt on. For example, if the version is 2.1.1, we would check for
2.1.0. There are a couple of problems with this:
* if the patch version is 0, we don't know what is the patch version
from the previous version
* this assume that the previous released was active a bit and generated
syncs

Note that the above are not mutually exclusive so we could have the
second option work with the solution in this PR so maybe this is another
way to alleviate the problem in the future.

You can see the parameter beings passed
[here](https://github.com/airbytehq/airbyte/actions/runs/18538584966/job/52840046885)
when added to the GitHub action and is not passed when not added
[here](https://github.com/airbytehq/airbyte/actions/runs/18540377411/job/52846048671)

## Review guide
<!--
1. `x.py`
2. `y.py`
-->

## User Impact
<!--
* What is the end result perceived by the user?
* If there are negative side effects, please list them. 
-->

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [x] YES 💚
- [ ] NO ❌