fix: publish latest tag for python connectors on main release (#68199)
## What
Because of this default setting [the logic to determine whether to push
the `latest` tag
](https://github.com/airbytehq/airbyte/blob/032869af01ee49196d174be4a1f0e6b826ab0dea/.github/actions/connector-image-build-push/action.yml#L148-L149)on
connector images was evaluating to "false" even for "main releases"
(i.e. merging to master with a non pre-release or RC tag). This is
unintended behavior.

Here's an
[example](https://github.com/airbytehq/airbyte/actions/runs/18568223627/job/52935068035#step:18:394)
of a connector release where is was working incorrectly.

Testing:
- I ran a manual main-release publish workflow
[here](https://github.com/airbytehq/airbyte/actions/runs/18660728140)
and looked at the logs to ensure the latest logic was evaluated
properly. The image version for the chosen connector already exists on
Dockerhub, so we wont attempt to republish.
- I ran a manual pre-release publish workflow
[here](https://github.com/airbytehq/airbyte/actions/runs/18660880512)
and made sure no latest-tag was added

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [x] YES 💚
- [ ] NO ❌