Safe debezium engine shutdown (#68208)
## What
While the engine is shutting down we don't accept any incoming event as:
a. It's incorrect since we already decided to close the engine (i.e
target reached)
b. It's unsafe for socket write as the engine forcefully kills it event
thread, which may cause our socket to disconnect if in the middle to
write

## How
Added a flag to signal engine shutting down and prevent accepting any
incoming event.

## Review guide
<!--
1. `x.py`
2. `y.py`
-->

## User Impact
<!--
* What is the end result perceived by the user?
* If there are negative side effects, please list them. 
-->

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [ ] YES 💚
- [ ] NO ❌